Add missing lvalue conversions
typeintpromote and typeargpromote assume they are dealing with an
unqualified type, and return an incorrect result if they are given a
qualified one. So, add an assert here.

This was causing const integer types to get promoted to themselves due
to missing lvalue conversions.

Thanks to Andrew Chambers for the bug report and test case.