Change $(objdir) instead of $(srcdir) for stage2 and stage3 builds
This is a bit simpler, and avoids issues with FreeBSD make when
including deps.mk (since the source files listed there are not relative
to $(srcdir)).

To do this, we need to make use of make(1) pattern substitutions, which
are not yet standard (though accepted for issue 8[0]).

Also, since the path of the object files seems to end up the resulting
executables, breaking byte-wise comparison, make sure to strip the
resulting executables.

[0] http://austingroupbugs.net/view.php?id=519