Revert "expr: Fix cast of same type"
This reverts commit 1a38a5fc4844a0de8729be694a62ba0afce3ff52.

This breaks comparisons bitfields in some cases, for instance

	extern struct {unsigned x:31;} s;
	int main(void) {
		return (unsigned)s.x - 1 < 0;
	}

If we discard the cast, then it is a signed comparison because of integer
promotion for bit-fields, otherwise it is an unsigned comparison.

Additionally, the test case this was meant to fix is not actually ISO C,
since casts must be to scalar types or `void`.