qbe: Remove unnecessary extension for relational operands
The operands are already promoted, so never need to be extended.

These extend()s were added in 7d746860 in an attempt to fix a bug
related to the comparison of values with type smaller than int.
However, the real bug was that the operands should have been promoted
to int by usual arithmetic conversions, which was fixed later in
a8131372.