Use architecture-specific va_list type
Previously, cproc effectively used used

	typedef struct { /* 32 bytes, 8-byte aligned */ } __builtin_va_list[1];

However, this is not quite correct for x86_64 nor aarch64, though
it was close enough for both to work in most cases.

In actuality, for x86_64 we want

	typedef struct { /* 24 bytes, 8-byte aligned */ } __builtin_va_list[1];

and for aarch64 we want

	typedef struct { /* 32 bytes, 8-byte aligned */ } __builtin_va_list;

The difference only appears when the size of va_list matters, or
when va_list is passed as a parameter. However, the former is not
often the case, and the aarch64 ABI replaces aggregate arguments
with pointers to caller-allocated memory, which is quite similar
to arrays decaying to pointers in C except that the struct is not
copied.

Additionally, riscv64 simply uses

	typedef void *__builtin_va_list;

which again has a different size and calling convention.

To fix this, make the __builtin_va_list type architecture-specific
and use architecture-specific tests for varargs-related functionality.