Fix type-checking of va_list arguments to varargs built-ins
If the argument was a function parameter, its type has already been
adjusted. So on x86_64, we can't just ignore the automatic
array-to-pointer conversion, since it was never a pointer to begin
with.

Instead, keep track of the adjusted va_list type, and check that
the arguments to varargs built-ins match that type.