qbe: Only return 0 from main if it has type int
Though C11 5.1.2.2.1 says that main must have a return type of int,
we could still encounter a program which declares it as something
else. This is undefined behavior, but we should not produce invalid
QBE IL in this case.

Also, 5.1.2.2.3 specifies that the implicit return 0 should only
apply when main's return type is compatible with int.