specialized branch compilation
This patch is a minor refactor of code generation for
branch conditions. Until now branch conditions were
treated just like any other expression. With the
present patch, and its associated modest increase in
complexity, branch conditions are compiled in a
context where the two targets (taken / not taken) are
available. This leads to slightly simpler qbe IL.

cproc bootstraps fine with the patch. Coremark sees
a 12% improvement when using qbe 1.2 as backend.