qbe: Avoid redundant leaf ceq in switch (v2)
Thread existing checked ranges through casesearch() in order
to avoid generating redundant comparisons/blocks for unit
range.

One minor improvement to previous patch - treat 0 case
specially (we could do the same at the max val side but that
seems unlikely to trigger in real code).

Previously posted that errors were encountered with v1 of
the patch; however I believe this was due to the bad 'TLNOT'
patch, and not this one.