Add defineobj() function and use when we define an object
This gives us a place to update alignment based on the final type,
which is relevant for external definitions and tentative definitions.
This fixes a bug where an object is declared with an incomplete
struct type, the struct type is completed, and then the object is
defined. This would previously result in an object declared with 0
alignment.

Also, check for incomplete types when we are defining objects.
Although the wording isn't clear in the C standard, and it doesn't
seem to disallow a thread_local definition of an object with
incomplete type, this doesn't make sense.