feat: Add -fno-builtin flag that does nothing
For gcc, -fno-builtin replaces built-in standard library functions with
gcc builtins that are faster. Some configure scripts for popular
packages test for this. However, cproc does not do this replacement. So,
we just make this flag do nothing.

Since there are now a few flags that do nothing but exist for
compatibility with other C compilers, I've added a mechanism to group
all of these ignored flags into one mechanism.

Signed-off-by: John Nunley <dev@notgull.net>